Using general tool MockLdap in AutoBDIISEs.py
